refactor(apm): Break APM commands out to specific files
